feat: update package dependencies and add stats command
- Removed unused dependencies from package.json.
- Updated lefthook version to 2.0.13.
- Added new stats command to display user ranking statistics.
- Refactored ranking command to simplify leaderboard display.
- Enhanced match embed creation with improved formatting and sorting.
- Improved error handling and user feedback in voting process.
