+++ package.json
@@ -19,9 +19,7 @@
 	},
 	"dependencies": {
 		"@hexcuit/server": "0.9.0",
-		"@resvg/resvg-js": "2.6.2",
-		"discord.js": "14.25.1",
-		"satori": "0.18.3"
+		"discord.js": "14.25.1"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "2.3.10",
@@ -29,7 +27,7 @@
 		"@changesets/cli": "2.29.8",
 		"@types/bun": "1.3.5",
 		"@types/react": "19.2.7",
-		"lefthook": "2.0.12",
+		"lefthook": "2.0.13",
 		"typescript": "5.9.3"
 	}
 }
