+++ src/commands/index.ts
@@ -5,4 +5,5 @@ export { default as admin } from './admin'
 export { default as queue } from './queue'
 export { default as ranking } from './ranking'
 export { default as register } from './register'
+export { default as stats } from './stats'
 export { default as team } from './team'
